Block Editor: Extract searchItems into its own file
