+++ packages/block-editor/src/components/inserter/test/menu.js
@@ -8,91 +8,12 @@ import ReactDOM from 'react-dom';
 /**
  * Internal dependencies
  */
-import { InserterMenu, searchItems, normalizeTerm } from '../menu';
-
-const textItem = {
-	id: 'core/text-block',
-	name: 'core/text-block',
-	initialAttributes: {},
-	title: 'Text',
-	category: 'common',
-	isDisabled: false,
-	utility: 1,
-};
-
-const advancedTextItem = {
-	id: 'core/advanced-text-block',
-	name: 'core/advanced-text-block',
-	initialAttributes: {},
-	title: 'Advanced Text',
-	category: 'common',
-	isDisabled: false,
-	utility: 1,
-};
-
-const someOtherItem = {
-	id: 'core/some-other-block',
-	name: 'core/some-other-block',
-	initialAttributes: {},
-	title: 'Some Other Block',
-	category: 'common',
-	isDisabled: false,
-	utility: 1,
-};
-
-const moreItem = {
-	id: 'core/more-block',
-	name: 'core/more-block',
-	initialAttributes: {},
-	title: 'More',
-	category: 'layout',
-	isDisabled: true,
-	utility: 0,
-};
-
-const youtubeItem = {
-	id: 'core-embed/youtube',
-	name: 'core-embed/youtube',
-	initialAttributes: {},
-	title: 'YouTube',
-	category: 'embed',
-	keywords: [ 'google' ],
-	isDisabled: false,
-	utility: 0,
-};
-
-const textEmbedItem = {
-	id: 'core-embed/a-text-embed',
-	name: 'core-embed/a-text-embed',
-	initialAttributes: {},
-	title: 'A Text Embed',
-	category: 'embed',
-	isDisabled: false,
-	utility: 0,
-};
-
-const reusableItem = {
-	id: 'core/block/123',
-	name: 'core/block',
-	initialAttributes: { ref: 123 },
-	title: 'My reusable block',
-	category: 'reusable',
-	isDisabled: false,
-	utility: 0,
-};
-
-const items = [
-	textItem,
-	advancedTextItem,
-	someOtherItem,
-	moreItem,
-	youtubeItem,
-	textEmbedItem,
-	reusableItem,
-];
+import items, { categories } from './fixtures';
+import { InserterMenu } from '../menu';
 
 const DEFAULT_PROPS = {
 	position: 'top center',
+	categories,
 	items,
 	debouncedSpeak: noop,
 	fetchReusableBlocks: noop,
@@ -321,51 +242,5 @@ describe( 'InserterMenu', () => {
 		expect( visibleBlocks[ 2 ].textContent ).toBe( 'A Text Embed' );
 
 		assertNoResultsMessageNotToBePresent( element );
-	} );
-} );
-
-describe( 'searchItems', () => {
-	it( 'should search items using the title ignoring case', () => {
-		expect( searchItems( items, 'TEXT' ) ).toEqual(
-			[ textItem, advancedTextItem, textEmbedItem ]
-		);
-	} );
-
-	it( 'should search items using the keywords', () => {
-		expect( searchItems( items, 'GOOGL' ) ).toEqual(
-			[ youtubeItem ]
-		);
-	} );
-
-	it( 'should search items using the categories', () => {
-		expect( searchItems( items, 'LAYOUT' ) ).toEqual(
-			[ moreItem ]
-		);
-	} );
-
-	it( 'should ignore a leading slash on a search term', () => {
-		expect( searchItems( items, '/GOOGL' ) ).toEqual(
-			[ youtubeItem ]
-		);
-	} );
-} );
-
-describe( 'normalizeTerm', () => {
-	it( 'should remove diacritics', () => {
-		expect( normalizeTerm( 'média' ) ).toEqual(
-			'media'
-		);
-	} );
-
-	it( 'should trim whitespace', () => {
-		expect( normalizeTerm( '  média  ' ) ).toEqual(
-			'media'
-		);
-	} );
-
-	it( 'should convert to lowercase', () => {
-		expect( normalizeTerm( '  Média  ' ) ).toEqual(
-			'media'
-		);
 	} );
 } );
